Refactor HeapSizeManager
- Rename to RpcThrottler
- Delegate heap size / rpc count limiting to single place
